#191 Remoção do pacote enumerador e movendo os enumeradores para o pacote model
