+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/HorarioAtendimento.java
@@ -5,7 +5,6 @@
  */
 package br.edu.ifpe.garanhuns.sg.model;
 
-import br.edu.ifpe.garanhuns.sg.model.enumarador.DiasSemana;
 import java.io.Serializable;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
