+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/dao/ConsultaDAO.java
@@ -9,8 +9,8 @@ import java.util.List;
 import br.edu.ifpe.garanhuns.sg.model.Consulta;
 import br.edu.ifpe.garanhuns.sg.model.Paciente;
 import br.edu.ifpe.garanhuns.sg.model.PostoSaude;
-import br.edu.ifpe.garanhuns.sg.model.enumarador.Especialidade;
-import br.edu.ifpe.garanhuns.sg.model.enumarador.Status;
+import br.edu.ifpe.garanhuns.sg.model.Especialidade;
+import br.edu.ifpe.garanhuns.sg.model.Status;
 import java.time.LocalDate;
 
 /**
