+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/dao/HorarioAtendimentoDAO.java
@@ -7,7 +7,7 @@ package br.edu.ifpe.garanhuns.sg.model.dao;
 
 import br.edu.ifpe.garanhuns.sg.model.HorarioAtendimento;
 import br.edu.ifpe.garanhuns.sg.model.PostoSaude;
-import br.edu.ifpe.garanhuns.sg.model.enumarador.Especialidade;
+import br.edu.ifpe.garanhuns.sg.model.Especialidade;
 import java.util.List;
 
 /**
