+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/dao/hibernate/AtendimentoHibernate.java
@@ -9,7 +9,7 @@ import java.util.List;
 import br.edu.ifpe.garanhuns.sg.model.dao.AtendimentoDAO;
 import br.edu.ifpe.garanhuns.sg.model.Atendimento;
 import br.edu.ifpe.garanhuns.sg.model.PostoSaude;
-import br.edu.ifpe.garanhuns.sg.model.enumarador.Especialidade;
+import br.edu.ifpe.garanhuns.sg.model.Especialidade;
 import org.hibernate.Session;
 import br.edu.ifpe.garanhuns.sg.util.HibernateUtil;
 import java.util.ArrayList;
