+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/dao/hibernate/ConsultaHibernate.java
@@ -12,8 +12,8 @@ import br.edu.ifpe.garanhuns.sg.model.Consulta;
 import br.edu.ifpe.garanhuns.sg.model.HorarioAtendimento;
 import br.edu.ifpe.garanhuns.sg.model.Paciente;
 import br.edu.ifpe.garanhuns.sg.model.PostoSaude;
-import br.edu.ifpe.garanhuns.sg.model.enumarador.Especialidade;
-import br.edu.ifpe.garanhuns.sg.model.enumarador.Status;
+import br.edu.ifpe.garanhuns.sg.model.Especialidade;
+import br.edu.ifpe.garanhuns.sg.model.Status;
 import org.hibernate.Session;
 import br.edu.ifpe.garanhuns.sg.util.HibernateUtil;
 import java.time.LocalDate;
