+++ src/main/java/br/edu/ifpe/garanhuns/sg/model/dao/hibernate/HorarioAtendimentoHibernate.java
@@ -10,7 +10,7 @@ import java.util.List;
 import br.edu.ifpe.garanhuns.sg.model.dao.HorarioAtendimentoDAO;
 import br.edu.ifpe.garanhuns.sg.model.HorarioAtendimento;
 import br.edu.ifpe.garanhuns.sg.model.PostoSaude;
-import br.edu.ifpe.garanhuns.sg.model.enumarador.Especialidade;
+import br.edu.ifpe.garanhuns.sg.model.Especialidade;
 import org.hibernate.Session;
 import br.edu.ifpe.garanhuns.sg.util.HibernateUtil;
